+++ odoo/local-src/specific_payment_mode/models/__init__.py
@@ -1,4 +1,6 @@
 # -*- coding: utf-8 -*-
 
+from . import account_payment_term
 from . import depiltech_payment_mode
+from . import sale_config
 from . import sale_order